combine generate-scripts
